+++ src/string/util/coerce.js
@@ -1,0 +1,8 @@
+import { isString } from '../../util/typeChecks';
+
+export default function coerce(str) {
+  if (!isString(str)) {
+    str = String(str);
+  }
+  return str;
+}