+++ test/tests/string.js
@@ -171,7 +171,7 @@ namespace('String', function() {
 
   describeInstance('pad', function(pad) {
 
-    it('should pad even length string to exact length', () => {
+    it('should pad even length string to exact length', function() {
       assertEqual(pad('wasabi', 0), 'wasabi');
       assertEqual(pad('wasabi', 1), 'wasabi');
       assertEqual(pad('wasabi', 2), 'wasabi');
@@ -187,7 +187,7 @@ namespace('String', function() {
       assertEqual(pad('wasabi', 20), '       wasabi       ');
     });
 
-    it('should pad odd length string to target length + 1', () => {
+    it('should pad odd length string to target length + 1', function() {
       assertEqual(pad('hello', 0), 'hello');
       assertEqual(pad('hello', 1), 'hello');
       assertEqual(pad('hello', 2), 'hello');
@@ -203,7 +203,7 @@ namespace('String', function() {
       assertEqual(pad('hello', 20), '       hello       ');
     });
 
-    it('should pad with custom string', () => {
+    it('should pad with custom string', function() {
       assertEqual(pad('wasabi', 8, '"'), '"wasabi"');
       assertEqual(pad('wasabi', 8, ''), 'wasabi');
       assertEqual(pad('wasabi', 8, 's'), 'swasabis');
@@ -212,7 +212,7 @@ namespace('String', function() {
       assertEqual(pad('hello', 12, '-'), '---hello---');
     });
 
-    it('should pad with non-standard arguments', () => {
+    it('should pad with non-standard arguments', function() {
       assertEqual(pad('wasabi'), 'wasabi');
       assertEqual(pad('wasabi', undefined), 'wasabi');
       assertEqual(pad('wasabi', null), 'wasabi');
@@ -222,7 +222,7 @@ namespace('String', function() {
       assertEqual(pad('', true), '');
     });
 
-    it('should not throw equivalent errors to padStart/padEnd', () => {
+    it('should not throw equivalent errors to padStart/padEnd', function() {
       assertNoError(function(){ pad('wasabi', -1); });
       assertNoError(function(){ pad('wasabi', -Infinity); });
       assertError(function(){ pad('wasabi',  Infinity); });
@@ -233,7 +233,7 @@ namespace('String', function() {
   describeInstance('truncate', function(truncate) {
     var str = 'Gotta be an entire sentence.';
 
-    it('should truncate to a specific length', () => {
+    it('should truncate to a specific length', function() {
       assertEqual(truncate(str, 29), 'Gotta be an entire sentence.');
       assertEqual(truncate(str, 28), 'Gotta be an entire sentence.');
       assertEqual(truncate(str, 21), 'Gotta be an entire se...');
@@ -249,7 +249,7 @@ namespace('String', function() {
       assertEqual(truncate(str, 0), '...');
     });
 
-    it('should truncate from the left', () => {
+    it('should truncate from the left', function() {
       assertEqual(truncate(str, 21, 'left'), '...e an entire sentence.');
       assertEqual(truncate(str, 11, 'left'), '...e sentence.');
       assertEqual(truncate(str, 9, 'left'), '...sentence.');
@@ -260,7 +260,7 @@ namespace('String', function() {
       assertEqual(truncate(str, -100, 'left'), '...');
     });
 
-    it('should should truncate from the middle', () => {
+    it('should should truncate from the middle', function() {
       assertEqual(truncate(str, 21, 'middle'), 'Gotta be an... sentence.');
       assertEqual(truncate(str, 11, 'middle'), 'Gotta ...ence.');
       assertEqual(truncate(str, 4, 'middle'), 'Go...e.');
@@ -270,7 +270,7 @@ namespace('String', function() {
       assertEqual(truncate(str, -100, 'middle'), '...');
     });
 
-    it('should allow a custom ellipsis', () => {
+    it('should allow a custom ellipsis', function() {
       assertEqual(truncate('string to truncate', 10, 'right', '|'), 'string to |');
       assertEqual(truncate('string to truncate', 10, 'right', 0), 'string to 0');
       assertEqual(truncate(str, 28, 'left', '>>> '), 'Gotta be an entire sentence.');
@@ -283,7 +283,7 @@ namespace('String', function() {
       assertEqual(truncate('123456', 2), '12...');
     });
 
-    it('should handle irregular input', () => {
+    it('should handle irregular input', function() {
       assertEqual(truncate(500, 2), '50...');
       assertEqual(truncate('short sentence', -1), '...');
       assertEqual(truncate('short sentence', 8, 'bad input'), 'short se...');
@@ -299,7 +299,7 @@ namespace('String', function() {
   describeInstance('truncateOnWord', function(truncateOnWord) {
     var str = 'Gotta be an entire sentence.';
 
-    it('should truncate to a specific length', () => {
+    it('should truncate to a specific length', function() {
       assertEqual(truncateOnWord(str, 100), 'Gotta be an entire sentence.');
       assertEqual(truncateOnWord(str, 28), 'Gotta be an entire sentence.');
       assertEqual(truncateOnWord(str, 27), 'Gotta be an entire...');
@@ -319,7 +319,7 @@ namespace('String', function() {
       assertEqual(truncateOnWord(str,  0), '...');
     });
 
-    it('should truncate from left', () => {
+    it('should truncate from left', function() {
       assertEqual(truncateOnWord(str, 21, 'left'), '...an entire sentence.');
       assertEqual(truncateOnWord(str, 20, 'left'), '...an entire sentence.');
       assertEqual(truncateOnWord(str, 19, 'left'), '...an entire sentence.');
@@ -338,7 +338,7 @@ namespace('String', function() {
       assertEqual(truncateOnWord(str,  0, 'left'), '...');
     });
 
-    it('should truncate from the middle', () => {
+    it('should truncate from the middle', function() {
       assertEqual(truncateOnWord(str, 21, 'middle'), 'Gotta be...sentence.');
       assertEqual(truncateOnWord(str, 20, 'middle'), 'Gotta be...sentence.');
       assertEqual(truncateOnWord(str, 19, 'middle'), 'Gotta be...sentence.');
@@ -357,7 +357,7 @@ namespace('String', function() {
       assertEqual(truncateOnWord(str, 0, 'middle'), '...');
     });
 
-    it('should should break on non-space punctuation', () => {
+    it('should should break on non-space punctuation', function() {
       assertEqual(truncateOnWord('a,short,string', 8), 'a,short...');
       assertEqual(truncateOnWord('a|short|string', 8), 'a|short...');
       assertEqual(truncateOnWord('a?short?string', 8), 'a?short...');
@@ -365,21 +365,21 @@ namespace('String', function() {
       assertEqual(truncateOnWord('a¿short¿string', 8), 'a¿short...');
     });
 
-    it('should should break on non-standard whitespace', () => {
+    it('should should break on non-standard whitespace', function() {
       assertEqual(truncateOnWord('a　short　string', 8), 'a　short...');
     });
 
-    it('should handle special cases', () => {
+    it('should handle special cases', function() {
       assertEqual(truncateOnWord('GOTTA BE AN ENTIRE SENTENCE.', 21), 'GOTTA BE AN ENTIRE...');
       assertEqual(truncateOnWord('gotta. be. an. entire. sentence.', 17), 'gotta. be. an....');
     });
 
-    it('should handle non-latin scripts', () => {
+    it('should handle non-latin scripts', function() {
       assertEqual(truncateOnWord('한국어 도 이렇게 할 수 있어요?', 9), '한국어 도 이렇게...');
       assertEqual(truncateOnWord('文字列　の　全角　スペース', 12), '文字列　の　全角...');
     });
 
-    it('should handle irregular input', () => {
+    it('should handle irregular input', function() {
       assertEqual(truncateOnWord(500, 2), '50...');
       assertEqual(truncateOnWord('short sentence', -1), '...');
       assertEqual(truncateOnWord('short sentence', 8, 'bad input'), 'short...');
@@ -387,13 +387,183 @@ namespace('String', function() {
       assertEqual(truncateOnWord('short sentence', 8, 'right', ''), 'short');
       assertEqual(truncateOnWord('short sentence', 8, 'right', null), 'shortnull');
       assertEqual(truncateOnWord('short sentence', 8, 'right', undefined), 'short...');
-      assertError(function(){ truncate('word', '8'); });
+      assertError(function(){ truncateOnWord('word', '8'); });
     });
 
-    it('should handle issues', () => {
+    it('should handle issues', function() {
       // #311
       assertEqual(truncateOnWord('Alpha Beta Gamma Delta Epsilon', 20, 'middle', ''), 'Alpha BetaEpsilon');
     });
 
   });
+
+  describeInstance('underscore', function(underscore) {
+
+    it('should handle basic input', function() {
+      assertEqual(underscore('hopOnPop'), 'hop_on_pop');
+      assertEqual(underscore('HopOnPop'), 'hop_on_pop');
+      assertEqual(underscore('HOPONPOP'), 'hoponpop');
+      assertEqual(underscore('HOP-ON-POP'), 'hop_on_pop');
+      assertEqual(underscore('hop-on-pop'), 'hop_on_pop');
+      assertEqual(underscore('watch me fail'), 'watch_me_fail');
+      assertEqual(underscore('watch   me   fail'), 'watch_me_fail');
+      assertEqual(underscore('watch me fail-sad-face'), 'watch_me_fail_sad_face');
+      assertEqual(underscore('waTch me su-cCeed'), 'wa_tch_me_su_c_ceed');
+    });
+
+    it('should not affect existing underscores', function() {
+      assertEqual(underscore('_hop_on_pop_'), '_hop_on_pop_');
+    });
+
+    it('should handle irregular input', function() {
+      assertEqual(underscore(''), '');
+      assertEqual(underscore(null), 'null');
+      assertEqual(underscore(800), '800');
+    });
+
+  });
+
+  describeInstance('camelize', function(camelize) {
+
+    it('should handle basic input', function() {
+      assertEqual(camelize('hop-on-pop'), 'HopOnPop');
+      assertEqual(camelize('HOP-ON-POP'), 'HopOnPop');
+      assertEqual(camelize('hop_on_pop'), 'HopOnPop');
+      assertEqual(camelize('watch me fail'), 'WatchMeFail');
+      assertEqual(camelize('watch   me   fail'), 'WatchMeFail');
+      assertEqual(camelize('watch me fail-sad-face'), 'WatchMeFailSadFace');
+      assertEqual(camelize('waTch me su-cCeed'), 'WaTchMeSuCCeed');
+    });
+
+    it('should downcase first letter', function() {
+      assertEqual(camelize('hop-on-pop', false), 'hopOnPop');
+      assertEqual(camelize('HOP-ON-POP', false), 'hopOnPop');
+      assertEqual(camelize('hop_on_pop', false), 'hopOnPop');
+      assertEqual(camelize('watch me fail', false), 'watchMeFail');
+      assertEqual(camelize('watch me fail-sad-face', false), 'watchMeFailSadFace');
+      assertEqual(camelize('waTch me su-cCeed', false), 'waTchMeSuCCeed');
+    });
+
+    it('should handle irregular input', function() {
+      assertEqual(camelize(''), '');
+      assertEqual(camelize(null), 'Null');
+      assertEqual(camelize(800), '800');
+    });
+
+  });
+
+  describeInstance('dasherize', function(dasherize) {
+
+    it('should handle basic input', function() {
+      assertEqual(dasherize('hop_on_pop'), 'hop-on-pop', 'underscores');
+      assertEqual(dasherize('HOP_ON_POP'), 'hop-on-pop', 'capitals and underscores');
+      assertEqual(dasherize('hopOnPop'), 'hop-on-pop');
+      assertEqual(dasherize('watch me fail'), 'watch-me-fail');
+      assertEqual(dasherize('watch me fail_sad_face'), 'watch-me-fail-sad-face');
+      assertEqual(dasherize('waTch me su_cCeed'), 'wa-tch-me-su-c-ceed');
+      assertEqual(dasherize('aManAPlanACanalPanama'), 'a-man-a-plan-a-canal-panama');
+      assertEqual(dasherize('street'), 'street');
+      assertEqual(dasherize('street_address'), 'street-address');
+      assertEqual(dasherize('person_street_address'), 'person-street-address');
+    });
+
+    it('should handle irregular input', function() {
+      assertEqual(dasherize(''), '');
+      assertEqual(dasherize(null), 'null');
+      assertEqual(dasherize(800), '800');
+    });
+
+  });
+
+  describeInstance('titleize', function(titleize) {
+
+    it('should handle basic input', function() {
+        assertEqual(titleize('active_record'), 'Active Record');
+        assertEqual(titleize('ActiveRecord'), 'Active Record');
+        assertEqual(titleize('action web service'), 'Action Web Service');
+        assertEqual(titleize('Action Web Service'), 'Action Web Service');
+        assertEqual(titleize('Action web service'), 'Action Web Service');
+        assertEqual(titleize('actionwebservice'), 'Actionwebservice');
+        assertEqual(titleize('Actionwebservice'), 'Actionwebservice');
+        assertEqual(titleize("david's code"), "David's Code");
+        assertEqual(titleize("David's code"), "David's Code");
+        assertEqual(titleize("david's Code"), "David's Code");
+        assertEqual(titleize('man from the boondocks'), 'Man from the Boondocks');
+        assertEqual(titleize('x-men: the last stand'), 'X Men: The Last Stand');
+        assertEqual(titleize('i am a sentence. and so am i.'), 'I Am a Sentence. And so Am I.');
+        assertEqual(titleize('hello! and goodbye!'), 'Hello! And Goodbye!');
+        assertEqual(titleize('hello, and goodbye'), 'Hello, and Goodbye');
+        assertEqual(titleize('hello; and goodbye'), 'Hello; And Goodbye');
+        assertEqual(titleize("about 'you' and 'me'"), "About 'You' and 'Me'");
+        assertEqual(titleize('TheManWithoutAPast'), 'The Man Without a Past');
+        assertEqual(titleize('raiders_of_the_lost_ark'), 'Raiders of the Lost Ark');
+    });
+
+    it('should handle non-titleized words', function() {
+      assertEqual(titleize('the day of the jackal'), 'The Day of the Jackal');
+      assertEqual(titleize('what color is your parachute?'), 'What Color Is Your Parachute?');
+      assertEqual(titleize('a tale of two cities'), 'A Tale of Two Cities');
+      assertEqual(titleize('where am i going to'), 'Where Am I Going To');
+    });
+
+    it('should handle irregular input', function() {
+      assertEqual(titleize(''), '');
+      assertEqual(titleize(null), 'Null');
+      assertEqual(titleize(800), '800');
+    });
+
+  });
+
+  describeInstance('spacify', function(spacify) {
+
+    it('should should handle basic input', function() {
+      assertEqual(spacify('hopOnPop'), 'hop on pop');
+      assertEqual(spacify('HopOnPop'), 'hop on pop');
+      assertEqual(spacify('HOPONPOP'), 'hoponpop');
+      assertEqual(spacify('HOP-ON-POP'), 'hop on pop');
+      assertEqual(spacify('hop-on-pop'), 'hop on pop');
+      assertEqual(spacify('watch_me_fail'), 'watch me fail');
+      assertEqual(spacify('watch-meFail-sad-face'), 'watch me fail sad face');
+      assertEqual(spacify('waTch me su-cCeed'), 'wa tch me su c ceed');
+    });
+
+    it('should handle irregular input', function() {
+      assertEqual(spacify(''), '');
+      assertEqual(spacify(null), 'null');
+      assertEqual(spacify(800), '800');
+    });
+  });
+
+  describeInstance('parameterize', function(parameterize) {
+
+    it('should handle basic input', function() {
+      assertEqual(parameterize('Donald E. Knuth'), 'donald-e-knuth');
+      assertEqual(parameterize('Random text with *(bad)* characters'), 'random-text-with-bad-characters');
+      assertEqual(parameterize('Allow_Under_Scores'), 'allow_under_scores');
+      assertEqual(parameterize('Trailing bad characters!@#'), 'trailing-bad-characters');
+      assertEqual(parameterize('!@#Leading bad characters'), 'leading-bad-characters');
+      assertEqual(parameterize('Squeeze   separators'), 'squeeze-separators');
+      assertEqual(parameterize('Test with + sign'), 'test-with-sign');
+    });
+
+    it('should allow a custom separator', function() {
+      assertEqual(parameterize('Donald E. Knuth', '_'), 'donald_e_knuth');
+      assertEqual(parameterize('Random text with *(bad)* characters', '_'), 'random_text_with_bad_characters');
+      assertEqual(parameterize('With-some-dashes', '_'), 'with-some-dashes');
+      assertEqual(parameterize('Retain_underscore', '_'), 'retain_underscore');
+      assertEqual(parameterize('Trailing bad characters!@#', '_'), 'trailing_bad_characters');
+      assertEqual(parameterize('!@#Leading bad characters', '_'), 'leading_bad_characters');
+      assertEqual(parameterize('Squeeze   separators', '_'), 'squeeze_separators');
+      assertEqual(parameterize('Test with + sign', '_'), 'test_with_sign');
+    });
+
+    it('should handle irregular input', function() {
+      assertEqual(parameterize(''), '');
+      assertEqual(parameterize(null), 'null');
+      assertEqual(parameterize(800), '800');
+      assertEqual(parameterize('Foo Bar', 8), 'foo8bar');
+      assertEqual(parameterize('Foo \uDFFF Bar'), 'foo-bar');
+    });
+  });
+
 });